Added Create Category Changed Long to Integer

+++ src/main/java/com/digitalmenu/menuservice/dish/DishRepository.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DishRepository
-        extends JpaRepository<Dish, Long> {
+public interface DishRepository extends JpaRepository<Dish, Integer> {
     Optional<Dish> findDishByName(String name);
 }
